Update to Thunderdome V4.2

+++ libraries/CurieBLE/src/BLEDevice.h
@@ -665,7 +665,7 @@ protected:
                                              const void *data, 
                                              uint16_t length);
     friend uint8_t profile_descriptor_read_rsp_process(bt_conn_t *conn, 
-                                            int err,
+                                            uint8_t err,
                                             bt_gatt_read_params_t *params,
                                             const void *data, 
                                             uint16_t length);
